python-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-generativeai==0.8.3
pillow==11.0.0
requests==2.32.3
pydantic==2.10.3
python-multipart==0.0.20

